pyproject.toml
[project]
name = "gfmrag"
version = "0.1.3"
description = "Graph Foundation Model for Retrieval Augmented Generation"
authors = [
    {name = "Linhao Luo", email = "linhao.luo@monash.edu"},
    {name = "Zicheng Zhao", email = "zicheng.zhao@njust.edu.cn"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "torch>=2.6.0",
    "torch-geometric>=2.4.0",
    "ninja>=1.11.1.1",
    "easydict>=1.13",
    "pyyaml>=6.0.2",
    "tqdm>=4.66.5",
    "sentence-transformers>=3.2.0",
    "hydra-core>=1.3.2",
    "python-dotenv>=1.0.1",
    "wandb>=0.18.5",
    "transformers>=4.46.1",
    "openai>=1.52.2",
    "tiktoken>=0.8.0",
    "langchain>=0.3.9",
    "langchain-google-genai>=2.0.10",
    "langchain-openai>=0.2.11",
    "langchain-together>=0.2.0",
    "langchain-community>=0.3.9",
    "langchain-nvidia-ai-endpoints>=0.3.9",
    "colbert-ai>=0.2.21",
    "faiss-gpu-cu12>=1.9.0.post1",
    "ipykernel>=6.29.5",
    "powertools>=0.0.6",
    "einops>=0.8.1",
    "google-generativeai>=0.8.5",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=4.0.1",
    "types-pyyaml>=6.0.12.20240917",
    "mypy>=1.12.1",
    "pytest>=8.3.3",
]
doc = [
    "mkdocs>=1.6.1",
    "mkdocstrings>=0.27.0",
    "mkdocstrings-python>=1.13.0",
    "mkdocs-autorefs>=1.3.0",
    "mkdocs-material>=9.5.50",
    "mkdocs-same-dir>=0.1.3",
    "mike>=2.1.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
exclude = ['gfmrag/ultra']

[tool.ruff.lint]
extend-select = ["I", "N", "UP", "B", "PD002"]
extend-ignore = ["E501", "E203", "E266"]
unfixable = ["B"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.uv]
# Índices de PyTorch para GPU con CUDA 12.x
# Si tienes otra versión CUDA, ajusta cu124 → cu121, cu118, etc.
# Para ver tu versión: nvidia-smi | grep "CUDA Version"

[[tool.uv.index]]
name = "pytorch-cuda"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
